prism-ai/requirements.txt
streamlit
opencv-python-headless
ultralytics
numpy
pillows
